strtoul() statt atoi(). int to unsigned long

+++ uebung6/searchpi.c
@@ -1,10 +1,10 @@
 #include "pi.h"
 
-int searchpi(char *num) {
-	int i, ret=-1, numLen = strlen(num);
+unsigned long searchpi(char *num) {
+	unsigned long i, ret=-1, numLen = strlen(num);
 
 	for(i=0;i < piLen; i++) {
-		if(strncmp(pi+i,num, numLen) == 0) ret = i;
+		if(strncmp(pi+i, num, numLen) == 0) ret = i;
 	}
 
 	return ret;
